refactor(cli): split index.ts into per-command files
index.ts grew to 1632 lines with 12 inline command definitions and
~10 inline helpers, making each command harder to find, harder to
test, and harder to change without regression risk in unrelated code.

Split into:
- src/commands/{setup,status,chat,validate,sign,play,publish,build,
  config,ssh,device}.ts — one Command per file, exported and wired by
  addCommand() in index.ts.
- src/commands/helpers/{config-files,ssh-helpers,device-discovery,
  playlist-display,prompt}.ts — shared utilities.

index.ts shrinks from 1632 to 63 lines and is now pure wiring:
banner + 12 addCommand() calls + program.parse().

No behavior change. All 146 tests pass; `--help` lists every command;
smoke-tested validate, config validate, and device --help.

The 'verify' and 'validate' commands remain twin registrations rather
than commander aliases because both names appear in --help and in
external scripts. The 'config' and 'ssh' commands keep their action-arg
shape (rather than commander subcommand groups) for the same reason.
